+++ bin-src/lib/getDependentStoryFiles.test.ts
@@ -382,6 +382,7 @@ describe('getDependentStoryFiles', () => {
     const rawContext = getContext();
     const ctx = {
       ...rawContext,
+      // signifying the package.json file had dependency changes
       git: { ...rawContext.git, changedPackageManifests: ['src/package.json'] },
     };
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
@@ -422,6 +423,34 @@ describe('getDependentStoryFiles', () => {
       changedFiles,
       changedDependencies
     );
+    expect(ctx.turboSnap.bailReason).toBeUndefined();
+    expect(res).toEqual({
+      './src/foo.stories.js': ['src/foo.stories.js'],
+    });
+  });
+
+  it('does not bail on dependency changes to package manifest file when untraced', async () => {
+    const changedFiles = ['src/foo.stories.js', 'src/package.json', 'src/package-lock.json'];
+    const modules = [
+      {
+        id: './src/foo.stories.js',
+        name: './src/foo.stories.js',
+        reasons: [{ moduleName: CSF_GLOB }],
+      },
+      {
+        id: CSF_GLOB,
+        name: CSF_GLOB,
+        reasons: [{ moduleName: './.storybook/generated-stories-entry.js' }],
+      },
+    ];
+    const rawContext = getContext({ untraced: ['**/(package.json|package-lock.json|yarn.lock)'] });
+    const ctx = {
+      ...rawContext,
+      // signifying the package.json file had dependency changes
+      git: { ...rawContext.git, changedPackageManifests: ['src/package.json'] },
+    };
+    const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
+    expect(ctx.turboSnap.bailReason).toBeUndefined();
     expect(res).toEqual({
       './src/foo.stories.js': ['src/foo.stories.js'],
     });