+++ bin-src/types.ts
@@ -216,6 +216,7 @@ export interface Context {
     testCount: number;
     changeCount: number;
     errorCount: number;
+    interactionTestFailuresCount: number;
     inProgressCount?: number;
     autoAcceptChanges: boolean;
     wasLimited?: boolean;